runner/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memory-bench"
version = "0.1.0"
description = "Customizable agent memory benchmark runner"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "memory-bench-designer contributors" }]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
embedding = [
    "sentence-transformers>=2.7",
    "numpy>=1.26",
]
dev = [
    "pytest>=8.0",
]

[project.scripts]
memory-bench = "memory_bench.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
